refactor: start commenting native methods

+++ Interop.cs
@@ -1,9 +1,9 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 using System.Security;
-using System.Text;
 
 // ReSharper disable InconsistentNaming
+
 namespace WinApi
 {
     [StructLayout(LayoutKind.Sequential)]
@@ -43,26 +43,22 @@ namespace WinApi
         public Rectangle PaintRectangle;
         public bool ReservedInternalRestore;
         public bool ReservedInternalIncUpdate;
-        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 32)]
-        public byte[] ReservedInternalRgb;
+        [MarshalAs(UnmanagedType.ByValArray, SizeConst = 32)] public byte[] ReservedInternalRgb;
     }
 
     [StructLayout(LayoutKind.Sequential)]
     public struct WindowClass
     {
         public WindowClassStyle Style;
-        [MarshalAs(UnmanagedType.FunctionPtr)]
-        public WindowProc WindowProc;
+        [MarshalAs(UnmanagedType.FunctionPtr)] public WindowProc WindowProc;
         public int ClassExtraBytes;
         public int WindowExtraBytes;
         public IntPtr InstanceHandle;
         public IntPtr IconHandle;
         public IntPtr CursorHandle;
         public IntPtr BackgroundBrushHandle;
-        [MarshalAs(UnmanagedType.LPTStr)]
-        public string MenuName;
-        [MarshalAs(UnmanagedType.LPTStr)]
-        public string ClassName;
+        [MarshalAs(UnmanagedType.LPTStr)] public string MenuName;
+        [MarshalAs(UnmanagedType.LPTStr)] public string ClassName;
     }
 
     [StructLayout(LayoutKind.Sequential)]
@@ -70,8 +66,7 @@ namespace WinApi
     {
         public int Size;
         public WindowClassStyle Style;
-        [MarshalAs(UnmanagedType.FunctionPtr)]
-        public WindowProc WindowProc;
+        [MarshalAs(UnmanagedType.FunctionPtr)] public WindowProc WindowProc;
         public int ClassExtraBytes;
         public int WindowExtraBytes;
         public IntPtr InstanceHandle;
@@ -127,29 +122,218 @@ namespace WinApi
 
         public static void Initialize(ref WindowPlacement obj)
         {
-            obj.Length = (int)Marshal.SizeOf<WindowPlacement>();
+            obj.Length = Marshal.SizeOf<WindowPlacement>();
         }
     }
 
     public delegate IntPtr WindowProc(IntPtr hwnd, uint msg, IntPtr wParam, IntPtr lParam);
+
     public delegate bool EnumWindowsProc(IntPtr hWnd, IntPtr lParam);
 
     [SuppressUnmanagedCodeSecurity]
     public static partial class NativeMethods
     {
+        /// <summary>
+        ///     Dispatches a message to a window procedure. It is typically used to dispatch a message retrieved by the GetMessage
+        ///     function.
+        /// </summary>
+        /// <param name="lpMsg">
+        ///     [Type: const MSG*]
+        ///     A pointer to a structure that contains the message.
+        /// </param>
+        /// <returns>
+        ///     [Type: LRESULT]
+        ///     The return value specifies the value returned by the window procedure.Although its meaning depends on the message
+        ///     being dispatched, the return value generally is ignored.
+        /// </returns>
+        /// <remarks>
+        ///     The MSG structure must contain valid message values. If the lpmsg parameter points to a WM_TIMER message and the
+        ///     lParam parameter of the WM_TIMER message is not NULL, lParam points to a function that is called instead of the
+        ///     window procedure.
+        ///     Note that the application is responsible for retrieving and dispatching input messages to the dialog box.Most
+        ///     applications use the main message loop for this. However, to permit the user to move to and to select controls by
+        ///     using the keyboard, the application must call IsDialogMessage.For more information, see Dialog Box Keyboard
+        ///     Interface.
+        /// </remarks>
         [DllImport("user32.dll")]
         public static extern IntPtr DispatchMessage([In] ref Message lpMsg);
 
+        /// <summary>
+        ///     Translates virtual-key messages into character messages. The character messages are posted to the calling thread's
+        ///     message queue, to be read the next time the thread calls the GetMessage or PeekMessage function.
+        /// </summary>
+        /// <param name="lpMsg">
+        ///     [Type: const MSG*]
+        ///     A pointer to an MSG structure that contains message information retrieved from the calling thread's message queue
+        ///     by using the GetMessage or PeekMessage function.
+        /// </param>
+        /// <returns>
+        ///     [Type: BOOL]
+        ///     If the message is translated(that is, a character message is posted to the thread's message queue), the return
+        ///     value is nonzero. If the message is WM_KEYDOWN, WM_KEYUP, WM_SYSKEYDOWN, or WM_SYSKEYUP, the return value is
+        ///     nonzero, regardless of the translation. If the message is not translated (that is, a character message is not
+        ///     posted to the thread's message queue), the return value is zero.
+        /// </returns>
+        /// <remarks>
+        ///     The TranslateMessage function does not modify the message pointed to by the lpMsg parameter.
+        ///     WM_KEYDOWN and WM_KEYUP combinations produce a WM_CHAR or WM_DEADCHAR message.WM_SYSKEYDOWN and WM_SYSKEYUP
+        ///     combinations produce a WM_SYSCHAR or WM_SYSDEADCHAR message.
+        ///     TranslateMessage produces WM_CHAR messages only for keys that are mapped to ASCII characters by the keyboard
+        ///     driver.
+        ///     If applications process virtual-key messages for some other purpose, they should not call TranslateMessage.For
+        ///     instance, an application should not call TranslateMessage if the TranslateAccelerator function returns a nonzero
+        ///     value.Note that the application is responsible for retrieving and dispatching input messages to the dialog box.Most
+        ///     applications use the main message loop for this. However, to permit the user to move to and to select controls by
+        ///     using the keyboard, the application must call IsDialogMessage.For more information, see Dialog Box Keyboard
+        ///     Interface.
+        /// </remarks>
         [DllImport("user32.dll")]
         public static extern bool TranslateMessage([In] ref Message lpMsg);
 
+        /// <summary>
+        ///     Retrieves a message from the calling thread's message queue. The function dispatches incoming sent messages until a
+        ///     posted message is available for retrieval. Unlike GetMessage, the PeekMessage function does not wait for a message
+        ///     to be posted before returning.
+        /// </summary>
+        /// <param name="lpMsg">
+        ///     [Type: LPMSG]
+        ///     A pointer to an MSG structure that receives message information from the thread's message queue.
+        /// </param>
+        /// <param name="hwnd">
+        ///     [Type: HWND]
+        ///     A handle to the window whose messages are to be retrieved.The window must belong to the current thread.
+        ///     If hWnd is NULL, GetMessage retrieves messages for any window that belongs to the current thread, and any messages
+        ///     on the current thread's message queue whose hwnd value is NULL (see the MSG structure). Therefore if hWnd is NULL,
+        ///     both window messages and thread messages are processed.
+        ///     If hWnd is -1, GetMessage retrieves only messages on the current thread's message queue whose hwnd value is NULL,
+        ///     that is, thread messages as posted by PostMessage (when the hWnd parameter is NULL) or PostThreadMessage.
+        /// </param>
+        /// <param name="wMsgFilterMin">
+        ///     [Type: UINT] The integer value of the lowest message value to be retrieved.Use WM_KEYFIRST (0x0100) to specify the
+        ///     first keyboard message or WM_MOUSEFIRST(0x0200) to specify the first mouse message.
+        ///     Use WM_INPUT here and in wMsgFilterMax to specify only the WM_INPUT messages.
+        ///     If wMsgFilterMin and wMsgFilterMax are both zero, GetMessage returns all available messages (that is, no range
+        ///     filtering is performed).
+        /// </param>
+        /// <param name="wMsgFilterMax">
+        ///     [Type: UINT]
+        ///     The integer value of the highest message value to be retrieved.Use WM_KEYLAST to specify the last keyboard message
+        ///     or WM_MOUSELAST to specify the last mouse message.
+        ///     Use WM_INPUT here and in wMsgFilterMin to specify only the WM_INPUT messages.
+        ///     If wMsgFilterMin and wMsgFilterMax are both zero, GetMessage returns all available messages (that is, no range
+        ///     filtering is performed).
+        /// </param>
+        /// <returns>
+        ///     [Type: BOOL]
+        ///     If the function retrieves a message other than WM_QUIT, the return value is nonzero.
+        ///     If the function retrieves the WM_QUIT message, the return value is zero.
+        ///     If there is an error, the return value is -1. For example, the function fails if hWnd is an invalid window handle
+        ///     or lpMsg is an invalid pointer.To get extended error information, call GetLastError.
+        /// </returns>
+        /// <returns>
+        ///     An application typically uses the return value to determine whether to end the main message loop and exit the
+        ///     program.
+        ///     The GetMessage function retrieves messages associated with the window identified by the hWnd parameter or any of
+        ///     its children, as specified by the IsChild function, and within the range of message values given by the
+        ///     wMsgFilterMin and wMsgFilterMax parameters. Note that an application can only use the low word in the wMsgFilterMin
+        ///     and wMsgFilterMax parameters; the high word is reserved for the system.
+        ///     Note that GetMessage always retrieves WM_QUIT messages, no matter which values you specify for wMsgFilterMin and
+        ///     wMsgFilterMax.
+        ///     During this call, the system delivers pending, nonqueued messages, that is, messages sent to windows owned by the
+        ///     calling thread using the SendMessage, SendMessageCallback, SendMessageTimeout, or SendNotifyMessage function. Then
+        ///     the first queued message that matches the specified filter is retrieved. The system may also process internal
+        ///     events. If no filter is specified, messages are processed in the following order:
+        ///     Sent messages
+        ///     Posted messages
+        ///     Input (hardware) messages and system internal events
+        ///     Sent messages (again)
+        ///     WM_PAINT messages
+        ///     WM_TIMER messages
+        ///     To retrieve input messages before posted messages, use the wMsgFilterMin and wMsgFilterMax parameters.
+        ///     GetMessage does not remove WM_PAINT messages from the queue. The messages remain in the queue until processed.
+        ///     If a top-level window stops responding to messages for more than several seconds, the system considers the window
+        ///     to be not responding and replaces it with a ghost window that has the same z-order, location, size, and visual
+        ///     attributes. This allows the user to move it, resize it, or even close the application. However, these are the only
+        ///     actions available because the application is actually not responding. When in the debugger mode, the system does
+        ///     not generate a ghost window.
+        /// </returns>
         [DllImport("user32.dll")]
         public static extern int GetMessage(out Message lpMsg, IntPtr hwnd, uint wMsgFilterMin,
             uint wMsgFilterMax);
 
+        /// <summary>
+        ///     Enables you to produce special effects when showing or hiding windows. There are four types of animation: roll,
+        ///     slide, collapse or expand, and alpha-blended fade.
+        /// </summary>
+        /// <param name="hwnd">
+        ///     [Type: HWND]
+        ///     A handle to the window to animate.The calling thread must own this window.
+        /// </param>
+        /// <param name="dwTime">
+        ///     [Type: DWORD]
+        ///     The time it takes to play the animation, in milliseconds.Typically, an animation takes 200 milliseconds to play.
+        /// </param>
+        /// <param name="dwFlags">
+        ///     [Type: DWORD]
+        ///     The type of animation.This parameter can be one or more of the following values. Note that, by default, these flags
+        ///     take effect when showing a window. To take effect when hiding a window, use AW_HIDE and a logical OR operator with
+        ///     the appropriate flags.
+        /// </param>
+        /// <returns>
+        ///     [Type: BOOL]
+        ///     If the function succeeds, the return value is nonzero.
+        ///     If the function fails, the return value is zero. The function will fail in the following situations:
+        ///     If the window is already visible and you are trying to show the window.
+        ///     If the window is already hidden and you are trying to hide the window.
+        ///     If there is no direction specified for the slide or roll animation.
+        ///     When trying to animate a child window with AW_BLEND.
+        ///     If the thread does not own the window. Note that, in this case, AnimateWindow fails but GetLastError returns
+        ///     ERROR_SUCCESS.
+        ///     To get extended error information, call the GetLastError function.
+        /// </returns>
+        /// <remarks>
+        ///     To show or hide a window without special effects, use ShowWindow.
+        ///     When using slide or roll animation, you must specify the direction. It can be either AW_HOR_POSITIVE,
+        ///     AW_HOR_NEGATIVE, AW_VER_POSITIVE, or AW_VER_NEGATIVE.
+        ///     You can combine AW_HOR_POSITIVE or AW_HOR_NEGATIVE with AW_VER_POSITIVE or AW_VER_NEGATIVE to animate a window
+        ///     diagonally.
+        ///     The window procedures for the window and its child windows should handle any WM_PRINT or WM_PRINTCLIENT messages.
+        ///     Dialog boxes, controls, and common controls already handle WM_PRINTCLIENT. The default window procedure already
+        ///     handles WM_PRINT.
+        ///     If a child window is displayed partially clipped, when it is animated it will have holes where it is clipped.
+        ///     AnimateWindow supports RTL windows.
+        ///     Avoid animating a window that has a drop shadow because it produces visually distracting, jerky animations.
+        /// </remarks>
         [DllImport("user32.dll", SetLastError = true)]
         public static extern int AnimateWindow(IntPtr hwnd, int dwTime, AnimateWindowFlag dwFlags);
 
+        /// <summary>
+        ///     Retrieves a handle to the top-level window whose class name and window name match the specified strings. This
+        ///     function does not search child windows. This function does not perform a case-sensitive search.
+        ///     To search child windows, beginning with a specified child window, use the FindWindowEx function.
+        /// </summary>
+        /// <param name="lpClassName">
+        ///     [Type: LPCTSTR]
+        ///     The class name or a class atom created by a previous call to the RegisterClass or RegisterClassEx function. The
+        ///     atom must be in the low-order word of lpClassName; the high-order word must be zero.
+        ///     If lpClassName points to a string, it specifies the window class name. The class name can be any name registered
+        ///     with RegisterClass or RegisterClassEx, or any of the predefined control-class names.
+        ///     If lpClassName is NULL, it finds any window whose title matches the lpWindowName parameter.
+        /// </param>
+        /// <param name="lpWindowName">
+        ///     [Type: LPCTSTR]
+        ///     The window name (the window's title). If this parameter is NULL, all window names match.
+        /// </param>
+        /// <returns>
+        ///     [Type: HWND]
+        ///     If the function succeeds, the return value is a handle to the window that has the specified class name and window
+        ///     name.
+        ///     If the function fails, the return value is NULL. To get extended error information, call GetLastError.
+        /// </returns>
+        /// <remarks>
+        ///     If the lpWindowName parameter is not NULL, FindWindow calls the GetWindowText function to retrieve the window name
+        ///     for comparison. For a description of a potential problem that can arise, see the Remarks for GetWindowText.
+        /// </remarks>
         [DllImport("user32.dll", CharSet = CharSet.Auto)]
         internal static extern IntPtr FindWindow(string lpClassName, string lpWindowName);
